+++ source/extra/ustdio/uprintf.c
@@ -276,7 +276,7 @@ u_vfprintf(    UFILE        *f,
 
     /* convert from the default codepage to Unicode */
     pattern = ufmt_defaultCPToUnicode(patternSpecification,
-        strlen(patternSpecification));
+        (int32_t)strlen(patternSpecification));
     if(pattern == 0) {
         return 0;
     }
@@ -380,7 +380,7 @@ u_printf_string_handler(UFILE                 *stream,
 
     /* convert from the default codepage to Unicode */
     if (arg) {
-        s = ufmt_defaultCPToUnicode(arg, strlen(arg));
+        s = ufmt_defaultCPToUnicode(arg, (int32_t)strlen(arg));
     }
     else {
         s = gNullStr;
